Separate center into a separate isolated module
